requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
antlr4-python3-runtime==4.8
argon2-cffi==21.1.0
async-timeout==3.0.1
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
crayons==0.4.0
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
einops==0.3.2
entrypoints==0.3
filelock==3.0.12
fsspec==2021.8.1
ftfy==6.0.3
future==0.18.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.39.0
huggingface-hub==0.0.16
idna==3.2
imageio==2.9.0
imageio-ffmpeg==0.4.5
imgtag==0.1.0
ipykernel==6.3.1
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.4
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==7.0.2
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.1
kornia==0.5.10
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
multidict==5.1.0
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.3
numpy==1.21.2
oauthlib==3.1.1
omegaconf==2.1.1
packaging==21.0
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
piexif==1.1.3
Pillow==7.1.2
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.17.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyDeprecate==0.3.1
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-xmp-toolkit==2.0.1
pytorch-lightning==1.4.4
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.2.1
regex==2021.8.28
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.45
Send2Trash==1.8.0
six==1.16.0
stegano==0.9.9
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
terminado==0.11.1
testpath==0.5.0
tokenizers==0.10.3
torch==1.9.0
torchmetrics==0.5.0
torchvision==0.10.0
tornado==6.1
tqdm==4.62.2
traitlets==5.1.0
transformers==4.10.0
typing-extensions==3.10.0.2
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
yarl==1.6.3
